requirements.txt
gradio>=4.0.0
yfinance>=0.2.0
dimod>=0.12.0
dwave-neal>=0.6.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
lxml>=4.9.0
requests>=2.28.0
